+++ src/types/Objects.ts
@@ -3,62 +3,103 @@ import Type from "../lang/Type";
 import Types from "../util/Types";
 import Assertions from "../util/Assertions";
 
-const hasOwnProperty = Object.prototype.hasOwnProperty;
-
 /**
- *
+ * Object interface
  */
 export interface ObjectsProps {
     [key: string]: any
 }
 
 /**
- * A class which provides some operations on Object
+ * Defines to return boolean or void in the callback
+ * @export
+ * @interface ForEachCallback
+ * ForEach Callback interface
+ */
+export interface ForEachCallback {
+    (item: any, key?: string, obj?: ObjectsProps): boolean | void
+}
+
+/**
+ * Defines to return value in the callback
+ * @export
+ * @interface MapCallback
+ * Map Callback interface
+ */
+export interface MapCallback {
+    (item: any, key?: string, obj?: Object): any
+}
+/**
+ * A class which provides some operations on Object {@see ObjectsProps}
  * @export
  * @class Objects
  */
 export default class Objects {
     /**
      *
-     * @param value
+     * Checks the given Object (is exist or not) or (the given key is exist or not}
+     * @param src {T}
+     * @param key? {string}
      * @return {boolean}
      */
-    static has(value: Object): boolean {
-        return has(value) && Objects.getLength(value) > 0
+    static has<T extends ObjectsProps>(src: T, key?: string): boolean {
+        return has(src) && ( has(key) ? src.hasOwnProperty(key): Objects.getLength(src) > 0);
     }
     /**
-     * gets length of the keys which are defined in the given object.
-     * @param obj
+     * Gets length of the given object keys
+     * @param src
      * @returns {number}
      */
-    public static getLength(obj: Object): number {
+    public static getLength<T extends ObjectsProps>(src: T): number {
         let i = 0;
-        for(let key in  obj) {
-            if(Objects.hasProperty(obj, key)) i++;
+        if(!has(src)) return i;
+        for(let key in  src) {
+            if(src.hasOwnProperty(key)) i++;
         }
         return i;
     }
+
     /**
-     *
-     * Checks the given given Object has key or not.
-     * @param object
-     * @param key
-     * @returns {boolean}
+     * Removes value by the given key from the given object
+     * @param src {T}
+     * @param key {any}
+     * @return {T}
      */
-    public static hasProperty(object: Object, key: string) {
-        return hasOwnProperty.call(object, key);
+    public static remove<T extends ObjectsProps>(src: T, key: any): T {
+        if(!has(src)) return src;
+        delete src[key];
+        return src;
+    }
+
+    /**
+     * Removes value from the given object
+     * @param src {T}
+     * @param value {any}
+     * @return {T}
+     */
+    public static removeValue<T extends ObjectsProps>(src: T, value: any): T {
+        if(!has(src)) return src;
+        for(let key in  src) {
+            if(src.hasOwnProperty(key)) {
+                if (src[key] == value) {
+                    delete src[key];
+                }
+            }
+        }
+        return src;
     }
     /**
-     * Provides to navigate in the given Object and create an array from the result of the callback function.
+     * Provdes to navigate in the given Object and create an array from the result of the callback function.
      * if the result of the callback is empty then ignores it.
      * @param obj
      * @param callback
      * @returns {Array<T>}
      */
-    public static map<T>(obj: Object, callback: (item: any, key?: string, obj?: Object) => T): T[] {
+    public static map<T extends ObjectsProps, M extends ObjectsProps>(obj: Object, callback: MapCallback ): any[] {
         let result = [];
+        if(!obj) return result;
         for(let key in  obj) {
-            if(Objects.hasProperty(obj, key)) {
+            if(obj.hasOwnProperty(key)) {
                 let callbackResult = callback(obj[key], key, obj);
                 if(!Type.hasNot(callbackResult)) {
                     result[result.length] = callbackResult;
@@ -70,13 +111,14 @@ export default class Objects {
 
     /**
      * Provides to navigate in the given Object and can broken if return false from callback function.
-     * @param obj
+     * @param obj { T }
      * @param callback
      * @returns {boolean}
      */
-    public static forEach(obj: Object, callback: (item: any, key?: string, obj?: Object) => boolean | void): boolean {
+    public static forEach<T extends ObjectsProps>(obj: T, callback: ForEachCallback): boolean {
+        if(!has(obj)) return true;
         for(let key in  obj) {
-            if(hasOwnProperty.call(obj, key)) {
+            if(obj.hasOwnProperty(key)) {
                 let callbackResult = callback(obj[key], key, obj);
                 if(callbackResult === false) {
                     return false;
@@ -86,47 +128,50 @@ export default class Objects {
         return true;
     }
     /**
-     * gets keys of the given object as string[].
-     * @param obj
+     * Gets keys of the given object as string[].
+     * @param src {T}
      * @returns {string[]}
      */
-    public static getKeys(obj: Object): string[] {
+    public static getKeys<T extends ObjectsProps>(src: T): string[] {
         let keys: string[] = [];
-        for(let key in  obj) {
-            if(Objects.hasProperty(obj, key)) keys.push(key);
+        if(!has(src)) return keys;
+        for(let key in src) {
+            if(src.hasOwnProperty(key)) keys.push(key);
         }
         return keys;
     }
 
     /**
-     *
-     * @param parentProp
-     * @param key
-     * @param value
-     * @param keys
+     * Adds value by key or keys to the given source Object
+     * @param src {T}
+     * @param key {string}
+     * @param value {any}
+     * @param keys? {string[]}
+     * @return {T}
      */
-    public static addValue(parentProp: ObjectsProps, key: string, value: any, keys?: string[]) {
+    public static addValue<T extends ObjectsProps>(src: T, key: string, value: any, keys?: string[]): T {
         if(keys && keys.length > 0) {
-            let prop = parentProp[key];
-            if(!Assertions.isObject(prop)) prop = parentProp[key] = {};
+            let prop = src[key];
+            if(!Assertions.isObject(prop)) prop = src[key] = {};
             let childKey = keys[0];
             keys.shift();
             Objects.addValue(prop, childKey, value, keys);
             keys.unshift(childKey);
         } else {
-            parentProp[key] = value;
+            src[key] = value;
         }
+        return src;
     }
 
     /**
-     *
-     * @param parentProp
-     * @param key
-     * @param keys
-     * @returns {any}
+     * Gets value by the given key or keys from the given source object
+     * @param src {T}
+     * @param key {string}
+     * @param keys? {string[]}
+     * @returns {T}
      */
-    public static getValue(parentProp: ObjectsProps, key: string, keys?: string[]): any  {
-        let prop = parentProp[key];
+    public static getValue<T extends ObjectsProps>(src: T, key: string, keys?: string[]): T  {
+        let prop = src[key];
         if(prop && keys && keys.length > 0) {
             for(let i = 0 ; i < keys.length; i++) {
                 prop = prop[keys[i]];
@@ -137,38 +182,28 @@ export default class Objects {
     }
 
     /**
-     *
-     * @param parentProp
-     * @param key
-     * @param keys
-     * @returns {any}
+     * Clones the given source object
+     * @param src {T}
+     * @param ignoreList? {string[]}
+     * @return {T}
      */
-    public static getValueByKeys(parentProp: ObjectsProps, keys?: string[]): any  {
-        let prop = parentProp;
-        for(let i = 0 ; i < keys.length; i++) {
-            prop = prop[keys[i]];
-            if(!prop) return null;
-        }
-        return prop;
-    }
-
-    public static clone(obj: Object, ignoreList?: string[]): Object {
-        return Types.getClone(obj, ignoreList);
+    public static clone<T extends ObjectsProps>(src: T, ignoreList?: string[]): T {
+        return Types.getClone(src, ignoreList);
     }
 
     /**
-     * Provides to merge src and destination by ignoreList.
-     * @param src
-     * @param dest
-     * @param ignoreList specifies get clone of the element by type or not clones just return its.
+     * Merges the given source object and destination object by ignoreList.
+     * @param src {S}
+     * @param dest {D}
+     * @param ignoreList? {string[]} specifies get clone of the element by type or not clones just return its.
      * @returns {any}
      */
-    public static merge(src: any, dest: any, ignoreList?: string[]): any {
+    public static merge<S extends ObjectsProps, D extends ObjectsProps>(src: S, dest: D, ignoreList?: string[]): S | D {
         if (src == null) return dest;
         if (dest == null) return Objects.clone(src,ignoreList);
         if(Assertions.isObject(src)) {
             for (let key in src) {
-                if (Objects.hasProperty(src, key)) {
+                if (src.hasOwnProperty(key)) {
                     if (Assertions.isObject(src[key]) && Assertions.isObject(dest[key])) {
                         Objects.merge(src[key], dest[key])
                     } else{
@@ -180,12 +215,18 @@ export default class Objects {
         return dest;
     }
 
-    public static mergeDefaults(defaults: any, props: any) {
+    /**
+     * Merges the given default props and the given props
+     * @param defaults {S}
+     * @param props {P}
+     * @return {any}
+     */
+    public static mergeDefaults<D extends ObjectsProps, P extends ObjectsProps>(defaults: D, props: P): D | P {
         if (defaults == null) return props;
         if (props == null) return Objects.clone(defaults,[]);
         if(Assertions.isObject(defaults)) {
             for (let key in defaults) {
-                if (Objects.hasProperty(defaults, key)) {
+                if (defaults.hasOwnProperty(key)) {
                     if(props[key]) {
                         if(Assertions.isObject(defaults[key]) && Assertions.isObject(props[key])) {
                             props[key] = Objects.mergeDefaults(defaults[key], props[key]);