+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGE LOG
 
+#### v1.0.3 Developing...
+* Changed method names and added comments to the **types/Objects** class
+* Changed method names and added comments to the **types/Arrays** class
+* Changed method names and added comments to the **types/Strings** class
+
 #### v1.0.2
 * fixed document link problems.